move NavBar logout logic to hook
